feat(popover): address pr comments
