+++ src/components/Popover/tippyPlugins.test.ts
@@ -1,4 +1,4 @@
-import { hideOnEscPlugin } from './tippyPlugins';
+import { hideOnEscPlugin, addBackdrop } from './tippyPlugins';
 
 describe('tippyPlugins', () => {
   describe('hideOnEscPlugin', () => {
@@ -6,6 +6,15 @@ describe('tippyPlugins', () => {
       expect(hideOnEscPlugin).toStrictEqual({
         name: 'hideOnEsc',
         defaultValue: true,
+        fn: expect.any(Function),
+      });
+    });
+  });
+
+  describe('addBackdrop', () => {
+    it('should return plugin correctly', () => {
+      expect(addBackdrop).toStrictEqual({
+        name: 'addBackdrop',
         fn: expect.any(Function),
       });
     });
